fixes activeMask issue
git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@5242 43c1dc8c-776e-4df2-8ef1-6e829299be21

+++ src/chan_sccp.h
@@ -1048,7 +1048,7 @@ struct sccp_device {
 	struct {
 		softkey_modes *modes;										/*!< used softkeySet */
 		uint8_t size;											/*!< how many softkeysets are provided by modes */
-		uint8_t activeMask[16];										/*!< enabled softkeys mask */
+		uint32_t activeMask[16];									/*!< enabled softkeys mask */
 	} softKeyConfiguration;											/*!< SoftKeySet configuration */
 	
 	struct {
